chore: fix 62 cargo warnings (unused imports/vars, dead assignments, mut/parens)
Applied cargo fix + cargo clippy --fix, then manually removed dead assignments
(db_done, checked_buckets, remaining, samples, id_autogen initial false) and
unused imports (Read/Write in conn_socket.rs, BufRead in syscheck.rs).
Removed unnecessary mut (adlist search_by matcher, expire db_done/checked_buckets,
unix remaining) and superfluous parens in eviction.rs.
All changes are mechanical/compiler-confirmed safe; no logic altered.
PORT NOTEs added where assignments were removed from TODO-stub paths.

+++ crates/redis-core/src/eviction.rs
@@ -138,7 +138,7 @@ fn sample_allkeys_lru(db: &RedisDb, samples: usize) -> Option<RedisString> {
 fn lfu_effective_counter(obj: &RedisObject, log_factor: u32, decay_time: u32) -> u8 {
     let raw = obj.lru;
     let counter = (raw & 0xFF) as u8;
-    let last_decrement_minutes = ((raw >> 8) & 0xFFFF) as u32;
+    let last_decrement_minutes = (raw >> 8) & 0xFFFF;
 
     let now_minutes = now_minutes();
     let elapsed = now_minutes.wrapping_sub(last_decrement_minutes);
@@ -277,7 +277,7 @@ pub fn lfu_init(obj: &mut RedisObject) {
 pub fn lfu_update(obj: &mut RedisObject, log_factor: u32, decay_time: u32) {
     let raw = obj.lru;
     let counter = (raw & 0xFF) as u8;
-    let last_decrement_minutes = ((raw >> 8) & 0xFFFF) as u32;
+    let last_decrement_minutes = (raw >> 8) & 0xFFFF;
 
     let now = now_minutes();
     let elapsed = now.wrapping_sub(last_decrement_minutes);
